Update main.json
I fixed some typos, changed a few "title"s that were incorrect, and responded to a few of your comments, Stephen. I won't mess with the rest until we can see it in action, but this should be the updated version of this. I'm going to try pasting it into a pool.